% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVlnorm.R
\name{CVlnorm}
\alias{CVlnorm}
\title{CV of log-normal dist baed on omega parameters
CV based variance like provided in Nonmem's \code{OMEGA} metrix.}
\usage{
CVlnorm(omega)
}
\arguments{
\item{omega}{A variance as provided in diagonal om the Nonmem OMEGA matrix}
}
\value{
CV of the distribution (numeric)
}
\description{
CV of log-normal dist baed on omega parameters
CV based variance like provided in Nonmem's \code{OMEGA} metrix.
}
\details{
This is a very simple function. All it does is
\code{sqrt(exp(omega)-1)}.
}
