% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqlog.R
\name{seqlog}
\alias{seqlog}
\title{Log-scale equidistant sequences}
\usage{
seqlog(from, to, length.out)
}
\arguments{
\item{from}{start of sequence}

\item{to}{end of sequence}

\item{length.out}{length of sequence}
}
\description{
Useful for generating sequences to be plotted on log scale.
}
