% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVlnorm.R
\name{CVlnorm}
\alias{CVlnorm}
\title{Convert a Variance of a normal distribution to a CV of a
log-normal distribution}
\usage{
CVlnorm(omega)
}
\arguments{
\item{omega}{A variance of a normal distribution}
}
\description{
Convert a Variance of a normal distribution to a CV of a
log-normal distribution
}
